Split AuthAction into AuthN and AuthZ
