+++ actions/AuthNAction.js
@@ -11,10 +11,8 @@
  */
 
 const { Core } = require('@adobe/aio-sdk');
-const LaunchDarkly = require('@launchdarkly/node-server-sdk');
 const QueryStringAddon = require('wretch/addons/queryString');
 const wretch = require('./Network.js');
-const { checkForAdobeInternalUser } = require('./ActionUtils.js');
 
 const logger = Core.Logger('AuthNAction');
 
@@ -75,97 +73,10 @@ async function getAccessToken(params) {
   }
 }
 
-async function getImsProfile(endpoint, clientId, token) {
-  try {
-    const response = await wretch(`${endpoint}/ims/profile/v1`)
-      .addon(QueryStringAddon).query({
-        client_id: clientId,
-      })
-      .headers({
-        Authorization: `Bearer ${token}`,
-        'Content-Type': 'application/json',
-      })
-      .get()
-      .json();
-    return response;
-  } catch (error) {
-    logger.error(error);
-    return null;
-  }
-}
-
-async function checkForProductContext(profile, org, productContext) {
-  try {
-    if (Array.isArray(profile.projectedProductContext)) {
-      const filteredProductContext = profile.projectedProductContext
-        .filter((obj) => obj.prodCtx.serviceCode === productContext);
-
-      // For each entry in filteredProductContext check that
-      // there is at least one entry where imsOrg matches the owningEntity property
-      // otherwise, if no match, the user is not authorized
-      return filteredProductContext.some((obj) => obj.prodCtx.owningEntity === org);
-    }
-    return false;
-  } catch (error) {
-    logger.error(error);
-    return false;
-  }
-}
-
-async function checkForEarlyProductAccess(toggle, sdkKey, isInternal, org) {
-  const ldClient = LaunchDarkly.init(sdkKey);
-  const context = {
-    kind: 'user',
-    key: org,
-    imsOrgId: org,
-    internal: isInternal,
-  };
-
-  return new Promise((resolve, reject) => {
-    ldClient.once('ready', () => {
-      ldClient.variation(toggle, context, false, (err, showFeature) => {
-        if (err) {
-          logger.error(err);
-          reject(err);
-        } else {
-          resolve(showFeature);
-        }
-      });
-    });
-
-    ldClient.on('error', (err) => {
-      logger.error(err);
-      reject(err);
-    });
-  });
-}
-
-function asAuthNAction(action, shouldAuthorize = true) {
+function asAuthNAction(action) {
   return async (params) => {
-    const imsEndpoint = params.IMS_ENDPOINT;
-    const clientId = params.IMS_CLIENT_ID;
-    const productContext = params.IMS_PRODUCT_CONTEXT;
-    const earlyAccessToggle = params.FT_EARLY_ACCESS;
-    const ldSdkKey = params.LD_SDK_KEY;
-
     const imsOrg = getImsOrg(params);
     const accessToken = await getAccessToken(params);
-
-    if (shouldAuthorize) {
-      // Check that the profile has access to the product
-      const imsProfile = await getImsProfile(imsEndpoint, clientId, accessToken);
-      if (!imsProfile) {
-        throw new Error('Failed to fetch profile');
-      }
-
-      if (!await checkForProductContext(imsProfile, imsOrg, productContext)) {
-        const isInternalUser = checkForAdobeInternalUser(imsProfile);
-
-        if (!await checkForEarlyProductAccess(earlyAccessToggle, ldSdkKey, isInternalUser, imsOrg)) {
-          throw new Error('Profile does not have access to the product');
-        }
-      }
-    }
 
     return action({ ...params, imsOrg, accessToken });
   };
