+++ actions/feedback/index.js
@@ -11,6 +11,7 @@
  */
 const { asGenericAction } = require('../GenericAction.js');
 const { asAuthNAction } = require('../AuthNAction.js');
+const { asAuthZAction } = require('../AuthZAction.js');
 const { asFirefallAction } = require('../FirefallAction.js');
 
 async function main(params) {
@@ -18,4 +19,4 @@ async function main(params) {
   return firefallClient.feedback(queryId, sentiment);
 }
 
-exports.main = asGenericAction(asAuthNAction(asFirefallAction(main)));
+exports.main = asGenericAction(asAuthNAction(asAuthZAction(asFirefallAction(main))));
